+++ tests/helper_cls.py
@@ -1,8 +1,9 @@
 # Licensed under Apache License Version 2.0 - see LICENSE
 
 # Built-ins
-import operator
 import itertools
+import operator
+import sys
 
 # This module
 import iteration_utilities
@@ -88,6 +89,44 @@ def failingTIterator(offset=0, repeats=1):
         return itertools.chain([T(1) for _ in range(offset)], it)
     else:
         return it
+
+
+class FailLengthHint(object):
+    """Simple iterator that fails when length_hint is called on it."""
+    def __init__(self, it):
+        self.it = iter(it)
+
+    def __iter__(self):
+        return self
+
+    def __next__(self):
+        return next(self.it)
+
+    def __length_hint__(self):
+        raise ValueError("length_hint failed")
+
+
+class OverflowLengthHint(object):
+    """Simple iterator that allows to set a length_hint so that one can test
+    overflow in PyObject_LengthHint.
+
+    Should be used together with sys.maxsize so it works on 32bit and 64bit
+    builds.
+    """
+    maxsize = sys.maxsize
+
+    def __init__(self, it, length_hint):
+        self.it = iter(it)
+        self.lh = length_hint
+
+    def __iter__(self):
+        return self
+
+    def __next__(self):
+        return next(self.it)
+
+    def __length_hint__(self):
+        return self.lh
 
 
 if iteration_utilities.EQ_PY2: